web/requirements.txt
flask==3.0.2
psycopg==3.1.18
python-dotenv